Fixed FileNotFoundError
Img for predict isn't saved in a certain folder, but in the same folder as the main.py

+++ flask_ml_app/main.py
@@ -1,13 +1,14 @@
-from flask import Flask, url_for, request, render_template, jsonify
+from flask import Flask, request,url_for, render_template, jsonify
 import tensorflow as tf
 import numpy as np
 import os
 from PIL import Image
+from werkzeug.utils import secure_filename
 
 #Global variables (load model and folder name)
 app = Flask(__name__)
-model = tf.keras.models.load_model("flask_ml_app/ham10000_model_v_augment_xception")
-temp_img = "temp" #Temporarily save file just in case. Will be deleted after prediction gives result
+model = tf.keras.models.load_model("ham10000_model_v_augment_xception")
+temp_img = "/home/m2243f2157/flask_ml_app/temp" #Temporarily save file just in case. Will be deleted after prediction gives result
 
 list_res = ['Actinic keratoses', 
             'basal cell carcinoma ', 
@@ -23,8 +24,7 @@ def index():
 
 #Function to predict what disease is this
 @app.route("/predict", methods=['GET','POST'])
-def predict_img():
-    print(request.url)
+def predict():
     response = {
                     'filename' : '',
                     'penyakit' : '',
@@ -32,11 +32,16 @@ def predict_img():
                 }
     if(request.method == 'POST'):
         
-        image = request.files.get("imagefile",None)
-        path = os.path.join(temp_img, image.filename)
-        image.save(path)
-        response = model_predict(path)
-        os.remove(path)
+        image = request.files.get("image_file")
+        if(image):
+                
+            image_filename = secure_filename(image.filename)
+            path = image_filename
+            image.save(path)
+            response = model_predict(path)
+            os.remove(path)
+        else:
+            response['solusi'] = "Image tidak terkirim"
     return jsonify(response)
 
 def model_predict(img_path):
@@ -50,8 +55,9 @@ def model_predict(img_path):
                     'penyakit' : list_res[np.argmax(results)],
                     'solusi' : "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
                 }
+    print(response)
     return response
     
 if __name__=="__main__":
-    app.run(host="0.0.0.0", port=8080, debug=True)
+    app.run(host="127.0.0.1", port=8080)
     #print(model_predict("ISIC_0034321.jpg"))
